Comunicaçao completa beagle-discovery com controle na beagle testado

+++ main-board/prj-remote-controlled-flight/subsystems/common/Provant/inc/proVantProtocol.h
@@ -118,7 +118,7 @@ public:
 	void multwii_debug(float debug1,float  debug2,float  debug3,float debug4);
 	void multwii_sendstack();
 	void multwii2_sendEscdata(int rpm[2],float current[2],float voltage[2]);
-	void multwii2_rcNormalize(int channel[7]);
+	void multwii2_rcNormalize(int16_t channel[7]);
 	void multwii2_sendControldatain(float rpy[3], float drpy[3], float position[3], float velocity[3]);
 	void multwii2_sendControldataout(float servo[2], float escTorque[2], float escRpm[2]);
 };
